+++ example/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { SocketIOProvider } from "../lib/io";
-import { ServerSent } from "./ServerSent";
 import { SocketIo } from "./socketio/SocketIo";
+import { ServerSent } from "./sse/ServerSent";
 
 export const App = () => (
   <SocketIOProvider url="http://localhost:3000">
