+++ example/sse/ServerSent.tsx
